Update Gemfile.lock to fix security vulernablity
According to GitHub, "The yajl-ruby dependency defined in Gemfile.lock
has a known high severity security vulnerability in version range < 1.3.1
and should be updated."

We are using the yajl-ruby dependency because we are using react-jsx-sprockets,
and that gem is dependent on the most updated version of yajl-ruby. Simply
updating our version of yajl-ruby will resolve the issue.

The specific security vulenerablity we are dealing with here is
https://nvd.nist.gov/vuln/detail/CVE-2017-16516